Check if meta has fields before printing label
If a repeater block has only one image in it, $mmm returns as "Array" which is printed on the page as the repeater block's title. If we check that the meta has fields first, if it doesn't, it will just return empty and not print anything...better than printing "Array" at least.